adding changes

+++ ClassifierClient/client.hpp
@@ -4,7 +4,7 @@
 #include <memory>
 
 struct IClassifierClient{
-    virtual float ClassifyMessage(const std::string& message) = 0;
+    virtual double ClassifyMessage(const std::string& message) = 0;
 };
 
 struct IClassifierFactory{
